+++ test/unit/v1/test_utils.py
@@ -8,7 +8,7 @@
 #
 ######################################################################
 
-from .test_base import TestBase
+from ..test_base import TestBase
 
 from .deps import b2_url_encode, b2_url_decode, choose_part_ranges, format_and_scale_number, format_and_scale_fraction