+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * `b2sdk.v1.sync` refactored to reflect `b2sdk.sync` structure
+* Make `B2Api.get_bucket_by_id` return populated bucket objects in v2
 
 ## [1.7.0] - 2021-04-22